requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
pyenchant==1.6.8
